+++ tests/client/e2eTests/helpers/RoleHelper.js
@@ -3,6 +3,7 @@
 var util = require('util');
 var shortid = require('shortid');
 var e2e = require('@openveo/test').e2e;
+var permissionLoader = process.require('app/server/loaders/permissionLoader.js');
 var Helper = e2e.helpers.Helper;
 
 /**
@@ -78,4 +79,140 @@ RoleHelper.prototype.getUpdateExample = function() {
     name: 'Role example new name',
     permissions: ['perm3']
   };
+};
+
+/**
+ * Gets the translated labels of permissions groups.
+ *
+ * @async
+ * @method getPermissionsGroups
+ * @param {Object} dictionary The dictionary of translations to translate labels
+ * @return {Promise} A promise resolving with the list of translated group labels
+ */
+RoleHelper.prototype.getPermissionsGroups = function(dictionary) {
+  var self = this;
+
+  return browser.waitForAngular().then(function() {
+    return self.flow.execute(function() {
+      var plugins = process.api.getPlugins();
+      var entities = process.api.getCoreApi().getEntities();
+      var deferred = protractor.promise.defer();
+
+      // Get the full list of permissions ordered by groups
+      permissionLoader.buildPermissions(entities, plugins, function(error, permissions) {
+        if (error)
+          deferred.reject(error);
+        else {
+          var groups = [];
+
+          permissions.forEach(function(permission) {
+
+            // A group is a permission with a label and sub permissions
+            if (permission.permissions && permission.label)
+              groups.push(self.translate(permission.label, dictionary));
+
+          });
+
+          deferred.fulfill(groups);
+        }
+      });
+
+      return deferred.promise;
+    });
+  });
+};
+
+/**
+ * Gets the translated names of permissions within a group.
+ *
+ * @async
+ * @method getGroupPermissions
+ * @param {String} expectedGroup The group to fetch permissions from
+ * @param {Object} dictionary The dictionary of translations to translate names
+ * @return {Promise} A promise resolving with the list of translated permissions for the group
+ */
+RoleHelper.prototype.getGroupPermissions = function(expectedGroup, dictionary) {
+  var self = this;
+
+  return browser.waitForAngular().then(function() {
+    return self.flow.execute(function() {
+      var plugins = process.api.getPlugins();
+      var entities = process.api.getCoreApi().getEntities();
+      var deferred = protractor.promise.defer();
+
+      // Get the full list of permissions ordered by groups
+      permissionLoader.buildPermissions(entities, plugins, function(error, groups) {
+        if (error)
+          deferred.reject(error);
+        else {
+          var permissions = [];
+
+          for (var i = 0; i < groups.length; i++) {
+            var group = groups[i];
+            var label = self.translate(group.label, dictionary);
+
+            if (label === expectedGroup) {
+
+              // Found the group
+
+              // Get permissions
+              group.permissions.forEach(function(permission) {
+                permissions.push(self.translate(permission.name, dictionary));
+              });
+
+              break;
+            }
+          }
+
+          deferred.fulfill(permissions);
+        }
+      });
+
+      return deferred.promise;
+    });
+  });
+};
+
+/**
+ * Gets permissions descriptors (Objects with permission name and group).
+ *
+ * @async
+ * @method getPermissions
+ * @param {Object} dictionary The dictionary of translations to translate names of groups and permissions
+ * @return {Promise} A promise resolving with permissions descriptors
+ */
+RoleHelper.prototype.getPermissions = function(dictionary) {
+  var self = this;
+
+  return browser.waitForAngular().then(function() {
+    return self.flow.execute(function() {
+      var plugins = process.api.getPlugins();
+      var entities = process.api.getCoreApi().getEntities();
+      var deferred = protractor.promise.defer();
+
+      // Get the full list of permissions ordered by groups
+      permissionLoader.buildPermissions(entities, plugins, function(error, groups) {
+        if (error)
+          deferred.reject(error);
+        else {
+          var permissions = [];
+
+          groups.forEach(function(group) {
+            var groupName = self.translate(group.label, dictionary);
+
+            group.permissions.forEach(function(permission) {
+              permissions.push({
+                name: self.translate(permission.name, dictionary),
+                group: groupName
+              });
+            });
+          });
+
+          deferred.fulfill(permissions);
+        }
+      });
+
+      return deferred.promise;
+    });
+  });
 };